Add support for windows, make code arch specific
